+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 - Add ability to set the opacity of the scripting overlay.
 - Add ability to get/set map header properties and read tile pixel data via the API.
 - Add button to copy the full metatile label to the clipboard in the Tileset Editor.
+- Add option to not open the most recent project on launch.
+- Add color picker to palette editor for taking colors from the screen.
 
 ### Changed
 - If an object event is inanimate, it will always render using its first frame.
@@ -25,6 +27,7 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 - Metatiles are always rendered accurately with 3 layers, and the unused layer is not assumed to be transparent.
 - `object_event_graphics_info.h` can now be parsed correctly if it uses structs with attributes.
 - The selection is no longer reset when pasting events. The newly pasted events are selected instead.
+- Palette editor ui is updated a bit to allow hex and rgb value input.
 
 ### Fixed
 - Fix cursor tile outline not updating at the end of a dragged selection.
@@ -33,6 +36,7 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 - Fix selected space not updating while painting in Collision view.
 - Fix the map music dropdown being empty when importing a map from Advance Map.
 - Fix object events added by pasting ignoring the map event limit.
+- Fixed a bug where saving the tileset editor would reselect the main editor's first selected metatile.
 
 ## [4.5.0] - 2021-12-26
 ### Added